# docs/swagger.yaml
basePath: /
definitions:
  github_com_irawankilmer_lms_backend_internal_dto.LoginInput:
    properties:
      identifier:
        type: string
      password:
        type: string
    required:
    - identifier
    - password
    type: object
  github_com_irawankilmer_lms_backend_internal_dto.PasswordResetInput:
    properties:
      password:
        minLength: 8
        type: string
      token:
        type: string
    required:
    - password
    - token
    type: object
  github_com_irawankilmer_lms_backend_internal_dto.PasswordResetRequestInput:
    properties:
      email:
        type: string
    required:
    - email
    type: object
  github_com_irawankilmer_lms_backend_internal_models.User:
    properties:
      createdAt:
        type: string
      email:
        type: string
      emailVerified:
        type: string
      id:
        type: integer
      lastActivity:
        type: string
      lastLogin:
        type: string
      password:
        minLength: 8
        type: string
      updatedAt:
        type: string
      username:
        description: validasi custom(tidak boleh mengandung spasi)
        maxLength: 32
        minLength: 3
        type: string
    required:
    - email
    - password
    - username
    type: object
host: localhost:8080
info:
  contact: {}
  description: This is a sample server for Learning Management System.
  title: Learning Management System API
  version: "1.0"
paths:
  /login:
    post:
      consumes:
      - application/json
      description: Login user with username or email and password
      parameters:
      - description: Login Input
        in: body
        name: dto.loginInput
        required: true
        schema:
          $ref: '#/definitions/github_com_irawankilmer_lms_backend_internal_dto.LoginInput'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties: true
            type: object
        "400":
          description: Bad Request
          schema:
            additionalProperties: true
            type: object
        "401":
          description: Unauthorized
          schema:
            additionalProperties: true
            type: object
      summary: Login user
      tags:
      - auth
  /password-reset/request:
    post:
      consumes:
      - application/json
      description: Mengirimkan email untuk reset password
      parameters:
      - description: Email untuk reset password
        in: body
        name: email
        required: true
        schema:
          $ref: '#/definitions/github_com_irawankilmer_lms_backend_internal_dto.PasswordResetRequestInput'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties: true
            type: object
        "400":
          description: Bad Request
          schema:
            additionalProperties: true
            type: object
      summary: Request Password Reset
      tags:
      - auth
  /password-reset/reset:
    post:
      consumes:
      - application/json
      description: Reset password menggunakan token
      parameters:
      - description: Token dan Password baru
        in: body
        name: passwordReset
        required: true
        schema:
          $ref: '#/definitions/github_com_irawankilmer_lms_backend_internal_dto.PasswordResetInput'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties: true
            type: object
        "400":
          description: Bad Request
          schema:
            additionalProperties: true
            type: object
      summary: Reset Password
      tags:
      - auth
  /users:
    get:
      description: Retrieve a list of all users
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/github_com_irawankilmer_lms_backend_internal_models.User'
            type: array
        "500":
          description: Internal Server Error
          schema:
            additionalProperties: true
            type: object
      security:
      - ApiKeyAuth: []
      summary: Get all users
      tags:
      - users
    post:
      consumes:
      - application/json
      description: Create a new user with the provided information
      parameters:
      - description: User Data
        in: body
        name: user
        required: true
        schema:
          $ref: '#/definitions/github_com_irawankilmer_lms_backend_internal_models.User'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/github_com_irawankilmer_lms_backend_internal_models.User'
        "400":
          description: Bad Request
          schema:
            additionalProperties: true
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties: true
            type: object
      summary: Create a new user
      tags:
      - users
  /users/{id}:
    delete:
      description: Delete a user by their ID
      parameters:
      - description: User ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "204":
          description: No Content
        "400":
          description: Bad Request
          schema:
            additionalProperties: true
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties: true
            type: object
      summary: Delete a user
      tags:
      - users
    get:
      description: Retrieve a user by their ID
      parameters:
      - description: User ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/github_com_irawankilmer_lms_backend_internal_models.User'
        "400":
          description: Bad Request
          schema:
            additionalProperties: true
            type: object
        "404":
          description: Not Found
          schema:
            additionalProperties: true
            type: object
      summary: Get user by ID
      tags:
      - users
    put:
      consumes:
      - application/json
      description: Update a user's information by their ID
      parameters:
      - description: User ID
        in: path
        name: id
        required: true
        type: integer
      - description: User Data
        in: body
        name: user
        required: true
        schema:
          $ref: '#/definitions/github_com_irawankilmer_lms_backend_internal_models.User'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/github_com_irawankilmer_lms_backend_internal_models.User'
        "400":
          description: Bad Request
          schema:
            additionalProperties: true
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties: true
            type: object
      summary: Update a user
      tags:
      - users
securityDefinitions:
  ApiKeyAuth:
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"